refactor: introduce resolver
