--- src/consts.ts
@@ -1,1 +1,0 @@
-export const PROJECT_ROOT = __dirname;
